% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefixList.R
\name{build_prefix_lines}
\alias{build_prefix_lines}
\title{Build PREFIX lines from tokens and IRIs (pure)}
\usage{
build_prefix_lines(tokens, iris)
}
\arguments{
\item{tokens}{character vector of prefix tokens (will be normalized)}

\item{iris}{character vector of namespace IRIs (will be normalized)}
}
\value{
character vector like "PREFIX token: <iri/>"
}
\description{
Build PREFIX lines from tokens and IRIs (pure)
}
\keyword{internal}
