% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classEndpoint.R
\name{classEndpoint}
\alias{classEndpoint}
\title{Deprecated alias for classificationList}
\usage{
classEndpoint(endpoint = "ALL")
}
\arguments{
\item{endpoint}{Character. One of "CELLAR", "FAO", "ALL".}
}
\value{
Output of classificationList().
}
\description{
`classEndpoint()` is deprecated.  
Please use [classificationList()] instead.
}
\examples{
\dontrun{
classEndpoint("ALL")   # deprecated
classificationList("ALL")  # use this instead
}
}
