% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureData.R
\name{structureData}
\alias{structureData}
\title{Obtain the structure of the classification tables from CELLAR/FAO or both.}
\usage{
structureData(prefix, conceptScheme, endpoint, language)
}
\arguments{
\item{prefix}{}

\item{conceptScheme}{}

\item{language}{By default set to "en". Optional}
}
\value{
\code{structureData()} returns the structure of a classification table from CELLAR/FAO or both.
}
\description{
Obtain the structure of the classification tables from CELLAR/FAO or both.
}
