% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCsvParameter.R
\name{testCsvParameter}
\alias{testCsvParameter}
\title{Validation of output csv parameter}
\usage{
testCsvParameter(arg_name, arg_value)
}
\arguments{
\item{arg_name}{The name in error handling to describe the argument, 
the name will be used in the validation messages.}

\item{arg_value}{The input argument, a full csv path including file name and extension.}
}
\value{
validation or error messages
}
\description{
Validation of the CSV argument to make sure it is of the correct format
}
\examples{
{
    CSVout = NULL
    result <- testCsvParameter("CSV", CSVout)
    }
}
