% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCorrespondenceTable.R
\name{aggregateCorrespondenceTable}
\alias{aggregateCorrespondenceTable}
\title{Aggregate a correspondence table to higher hierarchical levels}
\usage{
aggregateCorrespondenceTable(AB, A, B, CSVout = NULL)
}
\arguments{
\item{AB}{A mandatory argument containing a correspondence table data frame with columns "Acode" and "Bcode" representing the correspondence between classifications A and B at the most granular level. This argument is mandatory}

\item{A}{A path to a CSV file containing source classification data with an Acode ALevel,ASuperior column. This argument is mandatory}

\item{B}{A path to a CSV file containing target classification data with a Bcode Blevel BSuperior column. This argument is mandatory}

\item{CSVout}{a character string providing the path where the aggregated correspondence table CSV file should be saved. If NULL, no CSV file is generated.}
}
\value{
A data frame representing the aggregated correspondence table.
}
\description{
Aggregate a correspondence table to higher hierarchical levels.
}
\examples{
# Use data from the folder extdata
AB <- (system.file("extdata", "ab_data.csv", package = "correspondenceTables"))
A <- (system.file("extdata", "a_data.csv", package = "correspondenceTables"))
B <- (system.file("extdata", "b_data.csv", package = "correspondenceTables"))


result <- aggregateCorrespondenceTable(AB = AB, A = A, B = B, CSVout = NULL)
print(result)

}
