% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseCorrespondenceTable.R
\name{analyseCorrespondenceTable}
\alias{analyseCorrespondenceTable}
\title{analyseCorrespondenceTable performs analysis on correspondence tables}
\usage{
analyseCorrespondenceTable(
  AB,
  A = NULL,
  formatA = NULL,
  B = NULL,
  formatB = NULL,
  CSVcorrespondenceInventory = NULL,
  CSVcorrespondenceAnalysis = NULL
)
}
\arguments{
\item{AB}{A data frame containing the correspondence table data with columns "Acode" and "Bcode".}

\item{A}{A data frame or a path to a CSV file containing source classification data with "Acode" column.}

\item{formatA}{A regular expression pattern to filter source classification data based on "Acode".}

\item{B}{A data frame or a path to a CSV file containing target classification data with "Bcode" column.}

\item{formatB}{A regular expression pattern to filter target classification data based on "Bcode".}
}
\value{
A list containing two data frames: Annexe_A and Annexe_B.
The `Annexe_A` contains statistics related to components, correspondence types, and source/target positions.
The `Annexe_B` contains statistics for each class in the correspondence table.
}
\description{
The `analyseCorrespondenceTable` function takes input correspondence tables (AB) and related data (A and B) to perform analysis and generate various statistics.
It checks the validity of the input data, identifies components, calculates correspondence types, and creates summary tables.
}
\examples{
# Load correspondence table data

  AB_path   <- system.file("extdata", "ExempleAnnexe.csv", package = "correspondenceTables")
  AB <- read.csv(AB_path)





# Perform analysis
result <- analyseCorrespondenceTable(AB ,A = NULL, formatA = NULL, B = NULL, formatB = NULL, CSVcorrespondenceInventory = NULL, CSVcorrespondenceAnalysis = NULL) 
print(result$Annexe_A)
print(result$Annexe_B)
}
