% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseCorrespondancetable.R
\name{analyseCorrespondenceTable}
\alias{analyseCorrespondenceTable}
\title{Perform analysis on correspondence tables}
\usage{
analyseCorrespondenceTable(
  AB,
  A = NULL,
  B = NULL,
  longestAcodeOnly = FALSE,
  longestBcodeOnly = FALSE,
  CSVcorrespondenceInventory = NULL,
  CSVcorrespondenceAnalysis = NULL
)
}
\arguments{
\item{AB}{A mandatory argument containing a CSV file provided by the user
with the correspondence table data. The table must contain at least the
columns \code{"Acode"} and \code{"Bcode"}.}

\item{A}{A path to a CSV file containing source classification data with
an \code{"Acode"} column. Can be \code{NULL} if no source classification
table is provided.}

\item{B}{A path to a CSV file containing target classification data with
a \code{"Bcode"} column. Can be \code{NULL} if no target classification
table is provided.}

\item{longestAcodeOnly}{Logical. If \code{TRUE}, source classification data
are filtered based on \code{"Acode"} retaining only the maximum length,
i.e. the lowest-level Acodes in the correspondence table.}

\item{longestBcodeOnly}{Logical. If \code{TRUE}, target classification data
are filtered based on \code{"Bcode"} retaining only the maximum length,
i.e. the lowest-level Bcodes in the correspondence table.}

\item{CSVcorrespondenceInventory}{Character string giving the path to a CSV
file where the correspondence inventory should be written. If \code{NULL},
no CSV is produced.}

\item{CSVcorrespondenceAnalysis}{Character string giving the path to a CSV
file where the correspondence analysis should be written. If \code{NULL},
no CSV is produced.}
}
\value{
A list containing two data frames: \code{Inventory} and \code{Analysis}.
  \code{Inventory} contains statistics related to components, correspondence
  types, and source/target positions. \code{Analysis} contains statistics for
  each class in the correspondence table.
}
\description{
Perform analysis on a correspondence table using its associated
classifications to generate various statistics. It checks the validity of the
input data, identifies components, calculates correspondence types, and
creates summary tables.
}
\examples{
# Use data from the folder extdata
result <- analyseCorrespondenceTable(
  AB = system.file("extdata", "ExempleAnnexe.csv", package = "correspondenceTables"),
  A  = NULL,
  B  = NULL,
  longestAcodeOnly          = FALSE,
  longestBcodeOnly          = FALSE,
  CSVcorrespondenceInventory = NULL,
  CSVcorrespondenceAnalysis  = NULL
)
print(result$Inventory)
print(result$Analysis)
}
