% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classEndpoint.R
\name{classEndpoint}
\alias{classEndpoint}
\title{Retrieve a list of classification tables in CELLAR/FAO or both.}
\usage{
classEndpoint(endpoint)
}
\arguments{
\item{endpoint}{}
}
\value{
\code{classEndpoint()} returns a table with information needed to retreive the classification table:
\itemize{
    \item Prefix name
    \item Conceptscheme
    \item URI
    \item Name
}
}
\description{
Retrieve a list of classification tables in CELLAR/FAO or both.
}
