% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classEndpoint.R
\name{classEndpoint}
\alias{classEndpoint}
\title{Retrieve a list of classification tables in CELLAR, FAO or both.}
\usage{
classEndpoint(endpoint)
}
\arguments{
\item{endpoint.}{A string of type character containing the endpoint where the table is stored. 
The valid values are \code{"CELLAR"}, \code{"FAO"} and \code{"ALL"} for both endpoints.}
}
\value{
\code{classEndpoint()} returns a table with information needed to retrieve the classification table:
\itemize{
    \item Prefix name: the  SPARQL instruction for a declaration of a namespace prefix
    \item Conceptscheme: taxonomy of the SKOS object to be retrieved
    \item URI: the URL from which the SPARQL query was retrieved
    \item Name: the name of the table retrieved
}
}
\description{
Retrieve a list of classification tables in CELLAR, FAO or both.
}
\examples{
{
    endpoint = "ALL"
    list_data = classEndpoint(endpoint)
    }
}
