% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionClassification.R
\name{correctionClassification}
\alias{correctionClassification}
\title{Retrieve classification table from CELLAR and FAO repositories.}
\usage{
correctionClassification(classification)
}
\arguments{
\item{classification}{it returns a dataframe with two columns corrected according to the classification of CELLAR & FAO.}
}
\value{
\code{correctionClassification()} returns a table with information needed to retrieve the classification table:
 \itemize{
   \item Classification Code name (e.g. nace2): the code of each object
   \item Classification Label:  corresponding name of each object
 }
}
\description{
The aim of this function is to provide a table showing the different codes and labels for each classification
}
\examples{
{
prefix = "nace2"
conceptScheme = "nace2"
endpoint = "CELLAR"
classification = retrieveClassificationTable(prefix, endpoint, conceptScheme, level="ALL")$ClassificationTable
correct_classification = correctionClassification(classification)
View(correct_classification)
}
}
