% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionClassification.R
\name{correctionClassification}
\alias{correctionClassification}
\title{retrieve a corrected classification table from the CELLAR and FAO repositories for quality control purposes.}
\usage{
correctionClassification(classification)
}
\arguments{
\item{classification}{it returns a dataframe with two columns corrected according to the classification of CELLAR & FAO.}
}
\value{
\code{correctionClassification()} returns a table with information needed to retrieve the classification table:
 \itemize{
   \item Classification Code name (e.g. nace2): the code of each object
   \item Classification Label:  corresponding name of each object
 }
}
\description{
The main objective of this function is to prepare the classification table for quality control by ensuring the accuracy and consistency of the codes and labels. It performs various corrections based on specific classification requirements
}
\examples{
{
prefix = "nace2"
conceptScheme = "nace2"
endpoint = "CELLAR"
classification = retrieveClassificationTable(prefix, endpoint, conceptScheme, level="ALL")$ClassificationTable
classification = classification[,c(1,2)]
colnames(classification)[1:2] = c("Code", "Label")
correct_classification = correctionClassification(classification)
#View(correct_classification)
}
}
