% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondenceList.R
\name{correspondenceList}
\alias{correspondenceList}
\title{Retrieve a list of correspondence tables in CELLAR/FAO or both.}
\usage{
correspondenceList(endpoint)
}
\arguments{
\item{endpoint}{}
}
\value{
\code{correspondenceList()} returns a list of the correspondence tables available with their ID and prefix name
}
\description{
Retrieve a list of correspondence tables in CELLAR/FAO or both.
}
