% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondenceList.R
\name{correspondenceList}
\alias{correspondenceList}
\title{provides an overview of all the available correspondence classification from CELLAR and FAO repository.}
\usage{
correspondenceList(endpoint)
}
\arguments{
\item{endpoint}{The SPARQL Endpoint. The valid values are \code{"CELLAR"}, \code{"FAO"} or \code{"ALL"} for both.}
}
\value{
\code{correspondenceList()} returns a list of the correspondence tables available with prefix name, ID, Source classification, 
Target classification, Table name and URI.
}
\description{
provides an overview of all the available correspondence classification from CELLAR and FAO repository.
}
\examples{
{
    corr_list = correspondenceList("ALL")
}
}
