% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefixList.R
\name{prefixList}
\alias{prefixList}
\title{Create a list of prefixes for both CELLAR and FAO repositories.}
\usage{
prefixList(endpoint, prefix = NULL)
}
\arguments{
\item{endpoint}{SPARQL endpoints provide a standardized way to access data sets, 
making it easier to retrieve specific information or perform complex queries on linked data.  
The valid values are \code{"CELLAR"} and \code{"FAO"}.}
}
\value{
\code{prefixList()} returns a list of prefixes to be used when defying the SPARQL query.
}
\description{
Create a list of prefixes to be used when defying the SPARQL query to retrieve the tables
}
\examples{
{
    endpoint = "CELLAR"
    prefix_list = prefixList(endpoint)
    }
}
