% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefixList.R
\name{prefixList}
\alias{prefixList}
\title{Create a list of prefixes for both CELLAR and FAO}
\usage{
prefixList(endpoint)
}
\arguments{
\item{endpoint}{}
}
\value{
\code{prefixList()} returns a list of prefixes
}
\description{
Create a list of prefixes
}
