% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefixList.R
\name{prefixList}
\alias{prefixList}
\title{Create a list of prefixes for both CELLAR and FAO}
\usage{
prefixList(endpoint)
}
\arguments{
\item{endpoint.}{A string of type character containing the endpoint where the table is stored. 
The valid values are \code{"CELLAR"} and \code{"FAO"}.}
}
\value{
\code{prefixList()} returns a list of prefixes to be used when defying the SPARQL query.
}
\description{
Create a list of prefixes to be used when defying the SPARQL query to retrieve the tables
}
\examples{
{
    endpoint = "CELLAR"
    prefix_lst = prefixList(endpoint)
    }
}
