% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveClassificationTable.R
\name{retrieveClassificationTable}
\alias{retrieveClassificationTable}
\title{Retrieve a classification tables from CELLAR/FAO or both.}
\usage{
retrieveClassificationTable(
  prefix,
  endpoint,
  conceptScheme,
  level = "ALL",
  language = "en"
)
}
\arguments{
\item{prefix}{}

\item{endpoint}{}

\item{conceptScheme}{}

\item{language}{By default set to "en". Optional}

\item{level.}{By default set to "ALL". Optional}
}
\value{
\code{retrieveClassificationTable()} returns a classification tables from CELLAR/FAO or both.
}
\description{
Retrieve a classification tables from CELLAR/FAO or both.
}
