% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveCorrespondenceTable..R
\name{retrieveCorrespondenceTable}
\alias{retrieveCorrespondenceTable}
\title{Retrieve a correspondence tables from CELLAR/FAO or both.}
\usage{
retrieveCorrespondenceTable(prefix, endpoint, ID_table, language = "en")
}
\arguments{
\item{prefix}{}

\item{endpoint}{}

\item{ID_table}{}

\item{language}{By default set to "en". Optional}
}
\value{
\code{retrieveCorrespondenceTable()} returns a classification tables from CELLAR/FAO or both.
}
\description{
Retrieve a correspondence tables from CELLAR/FAO or both.
}
