% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureData.R
\name{structureData}
\alias{structureData}
\title{Obtain the structure of the classification tables from CELLAR and FAO.}
\usage{
structureData(prefix, conceptScheme, endpoint, language = "en")
}
\arguments{
\item{prefix}{The  SPARQL instruction for a declaration of a namespace prefix. It can be found using the classEndpoint() function.}

\item{conceptScheme}{Taxonomy of the SKOS object to be retrieved. It can be found using the classEndpoint() function.}

\item{endpoint}{The SPARQL Endpoint}

\item{language}{Language of the table. By default is set to \code{"en"}. This is an optional argument.}
}
\value{
\code{structureData()} returns the structure of a classification table from CELLAR and FAO in form a table with the following colums:        
 \itemize{
    \item Concept_Scheme: taxonomy of the SKOS object to be retrieved
    \item Level: the levels of the objects in the collection 
    \item Depth: identify the hierarchy of each level
    \item Count: the number of objects retrieved in each level
}
}
\description{
Obtain the structure of the classification tables from CELLAR and FAO.
}
\examples{
{
    endpoint = "CELLAR"
    prefix = "nace2"
    conceptScheme = "nace2"
    language = "en"
    structure_dt = structureData(prefix, conceptScheme, endpoint, language)
    }
}
